pyproject.toml
[project]
name = "alchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mammoth>=1.11.0",
    "pdfminer-six>=20250506",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]

